Submit a cromwell batch job

This function submits a set of one or more inputs to cromwell. It is much more efficient than submitting a single job at a time. See https://github.com/broadinstitute/cromwell#post-apiworkflowsversionbatch for details.

cromwellBatch(wdlSource, workflowInputs, workflowOptions = NULL,
  timeout = 120, ...)

Arguments

wdlSource

Represents the https://software.broadinstitute.org/wdl/A string (character vector of length 1) or an upload_file object. See details below.

workflowInputs

A data.frame that will be coerced to a json array or a JSON string (as a character vector of length 1), or an upload_file object. See details below.

workflowOptions

A list, a JSON string (as a character vector of length 1, or an upload_file object. See details below.

timeout

The number of seconds to wait for a response. Batch jobs can take quite some time for cromwell to process, so this will typically need to be set to a large value to allow for a completed response.

...

passed directly to httr `POST` (for including `timeouts`, `handles`, etc.)

Value

If a timeout does not occur (this is pretty common....), then a list that contains the submission status.

Details

TODO details